requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
langchain_pinecone
langchain-community
google-generativeai
cohere
streamlit